requirements.txt
gradio
groq
python-dotenv